[Misc] Fix implicit precision loss warnings

+++ src/theia/math/find_polynomial_roots_jenkins_traub.cc
@@ -314,7 +314,7 @@ bool JenkinsTraubSolver::ExtractRoots() {
   // Remove any leading zeros of the polynomial.
   polynomial_ = RemoveLeadingZeros(polynomial_);
 
-  const int degree = polynomial_.size() - 1;
+  const int degree = static_cast<int>(polynomial_.size()) - 1;
 
   // Allocate the output roots.
   if (real_roots_ != NULL) {
@@ -701,7 +701,7 @@ void JenkinsTraubSolver::RemoveZeroRoots() {
 }
 
 bool JenkinsTraubSolver::SolveClosedFormPolynomial() {
-  const int degree = polynomial_.size() - 1;
+  const int degree = static_cast<int>(polynomial_.size()) - 1;
 
   // Is the polynomial constant?
   if (degree == 0) {
